refactor(stripe): move webhook mapping into transformers, thin trait methods
Relocate all Stripe IncomingWebhook event-type / object-reference / status /
content mapping out of the trait methods and into stripe/transformers.rs as
pub(crate) free functions (get_webhook_object_from_body, map_webhook_event_type,
get_webhook_reference, build_webhook_payment_response / _refund_response /
_dispute_response), matching the canonical connector pattern (cf. hyperswitch,
adyen). The IncomingWebhook methods are now thin decode-then-delegate wrappers.

Pure refactor — behaviour is byte-for-byte identical (verified: ParseEvent +
HandleEvent regression 21/21, plus a real Stripe sandbox payment -> webhook
shadow diff keyDiff {} and primary served live by UCS).
